+++ server/auctionapp/src/main/java/com/atlantbh/auctionapp/repository/ProductRepository.java
@@ -7,12 +7,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
     Page<ProductEntity> findAllByEndDateIsAfter(LocalDateTime time ,Pageable paging);
-    Page<ProductEntity> findAllByCategoryIdInAndStartPriceBetweenAndEndDateIsAfter(long[] categoryId, BigDecimal lowPrice, BigDecimal highPrice, LocalDateTime time, Pageable paging);
+    Page<ProductEntity> findAllByCategoryIdInAndStartPriceBetweenAndEndDateIsAfter(long[] categoryId, double lowPrice, double highPrice, LocalDateTime time, Pageable paging);
     ProductEntity findProductById(long id);
 
     @Query(value = "SELECT MAX(start_price), MIN(start_price)" +
